Do not throw exceptions from fileExists()
This commit changes the function call to `fs.statSync()` by passing an additional option, `throwIfNoEntry`, as `false`. This has the effect of not throwing an exception and instead returning `undefined`.

This simplifies the logic and, more importantly, it treats the function as a simple boolean function without side-effects, as that is how it is used and should be used.